Adjusted pod filter to be an interface method
